Finished filling Aggregates with data

+++ main.py
@@ -1,3 +1,4 @@
+import re
 import PyPDF2
 import urllib.error
 import urllib.request
@@ -7,6 +8,35 @@ import logging
 import json
 
 configuration_filename = 'configuration.json'
+
+
+class Aggregate():
+    def __init__(self, name: str):
+        self.name = name
+        self.data = {}
+
+    def update_data(self, indicator: str, period: str, value: float):
+        if indicator not in self.data:
+            self.data.update({indicator: {}})
+        if period in self.data[indicator]:
+            raise ValueError
+        self.data[indicator].update({period: value})
+
+    def __str__(self):
+        bits = []
+        bits.append(self.name)
+        bits.append('\n')
+        for indicator in self.data:
+            bits.append(' ' * 2)
+            bits.append(indicator)
+            bits.append('\n')
+            for value in sorted(self.data[indicator]):
+                bits.append(' ' * 4)
+                bits.append(value)
+                bits.append(' ')
+                bits.append(str(self.data[indicator][value]))
+                bits.append('\n')
+        return ''.join(bits)
 
 
 def ensure_path_exists(path):
@@ -31,25 +61,18 @@ def initialize_sources(logger):
         filename = '{:02}.pdf'.format(i)
         full_url = base_path + filename
         source_path = os.path.join(root, filename)
-        urllib.request.urlretrieve(full_url, source_path)
+        if os.path.exists(source_path):
+            continue
         try:
             urllib.request.urlretrieve(full_url, source_path)
         except urllib.error.HTTPError:
             downloaded = i - 1
-            logger.info('Stopped after downloading {} file{}.'.format(downloaded, '' if downloaded else's'))
+            logger.info('Stopped after downloading {} file{}.'.format(downloaded, '' if downloaded == 1 else 's'))
             break
         except:
             logger.error('Unhandled exception!')
             raise
     logger.info('Finished downloading data files.')
-
-
-class Aggregate():
-    def __init__(self):
-        self.data = {}
-
-    def append(self, course, indicator, date, value):
-        pass
 
 
 def extract_text(pdf_filename):
@@ -76,3 +99,35 @@ def make_logger(filename):
 if __name__ == '__main__':
     logger = make_logger('log.txt')
     initialize_sources(logger)
+    with open(configuration_filename) as configuration_handler:
+        configuration = json.load(configuration_handler)
+    sources_root = configuration['sources_root']
+    # A map from course name to aggregate object.
+    aggregates = {}
+    for pdf_file in os.listdir(sources_root):
+        strings = extract_text(os.path.join(sources_root, pdf_file))
+        indicator = ' '.join(strings[2].split()[1:])
+        # The number of periods is variable, so we have to count them.
+        periods = []
+        i = 4
+        period_re = re.compile('\d{4}/\d')
+        while period_re.match(strings[i]):
+            periods.append(strings[i])
+            i += 1
+        # From this point onwards we will match the course name and the evaluations.
+        # There may be empty strings at the end, so we check against those.
+        while i < len(strings) and strings[i]:
+            name = strings[i]
+            i += 1
+            if name not in aggregates:
+                aggregates.update({name: Aggregate(name)})
+            for period in periods:
+                try:
+                    value = float(strings[i].replace(',', '.'))
+                except ValueError:
+                    value = None
+                aggregates.get(name).update_data(indicator, period, value)
+                i += 1
+
+    for aggregate in aggregates.values():
+        print(str(aggregate))
